requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
Pillow>=10.0
python-decouple>=3.8
whitenoise>=6.6
gunicorn>=21.2
psycopg2-binary>=2.9